[WEB-SERVICE] PHP Apache Web Server & MySQL Server
Deleted .dockerignore file, it is not needed anymore since we now use docker-compose.

Created docker-compose.yml file. This file allows to launch a PHP Apache Server and a MySQL Server. They are able to communicate.

Moved and modified the Dockerfile to import the PHP Apache Server image and install mysqli (communication with MySQL Server).

Moved the Docker Cheatsheet from the WEB\README.md to the Git Hub repository's Wiki. It now explains how to use the docker-compose.yml file,
and what output to expect.

Created an index.php simple file. It connects to the MySQL Server through mysqli and displays the connection status response.

Deleted the static html file (not usefull at the moment).