[INTERNAL-SERVICES] Created ERP & the direction "computer"
Opened port 53 on UDP and TCP for the internal DNS Server and the internal DNS Resolver Server.

Created the ERP Service. It is almost a carbon copy of the PHP Apache Web Server (the public one).

Slightly modified the PHP Apache Web Server to differanciate the ERP Server and the PHP Apache Web Server.

Modified the docker-compose configuration file to reflect all those changes.